mysite/requirements.txt
Django==2.2.2
django-mptt==0.10.0
social-auth-app-django==3.1.0
